Removed duplicate decline dialog

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/WorkorderBundleDetailActivity.java
@@ -22,7 +22,7 @@ import com.fieldnation.service.data.workorder.WorkorderClient;
 import com.fieldnation.ui.AuthSimpleActivity;
 import com.fieldnation.ui.RefreshView;
 import com.fieldnation.ui.dialog.v2.AcceptBundleDialog;
-import com.fieldnation.ui.dialog.v2.DeclineDialog;
+import com.fieldnation.v2.ui.dialog.DeclineDialog;
 
 public class WorkorderBundleDetailActivity extends AuthSimpleActivity {
     private static final String TAG = "WorkorderBundleDetailActivity";
@@ -159,7 +159,7 @@ public class WorkorderBundleDetailActivity extends AuthSimpleActivity {
         public void onClick(View v) {
             DeclineDialog.show(App.get(), UID_DIALOG_DECLINE,
                     _woBundle.getWorkorder().length,
-                    _woBundle.getWorkorder()[0].getWorkorderId(),
+                    _woBundle.getWorkorder()[0].getWorkorderId().intValue(),
                     _woBundle.getWorkorder()[0].getCompanyId());
         }
     };
